Evaluate m/z pairs against raw data.

EvaluateCandidateListAgainstRawData will compare the result of function EvaluatePairsFromXCMSSet against raw data files.

Usage

EvaluateCandidateListAgainstRawData(
  x = NULL,
  tp = NULL,
  gr = NULL,
  dat = NULL,
  dmz = 0.025,
  drt = 1,
  dEcut = 1,
  Pcut = 0.01,
  Icut = 1000,
  method = c("APCI", "ESI")[1],
  rolp = c("non", "pos", "neg", "all")[2],
  smooth = 0
)

Arguments

x

Dataframe of results (output of EvaluatePairsFromXCMSet).

tp

Timepoint.

gr

group, e.g. different genotypes or concentrations.

dat

list of xcmsRaw's for deconvolution and plotting.

dmz

Allowed mass deviation in Da (for BPC extraction).

drt

Allowed rt deviation in seconds (for get extraction).

dEcut

Minimum required change in enrichment before a candidate ID is assigned.

Pcut

Maximum allowed P value before a candidate ID is assigned.

Icut

Minimum required median peak intensity before a candidate ID is assigned.

method

Either APCI or ESI. Choice will modify some internal parameters and checks performed.

rolp

RemoveOverLappingPeaks parameter, overlapping means from a deconvoluted spectrum where another peak was already evaluated.

smooth

Smoothing parameter passed to getMultipleBPC.

Value

A list of evaluation results.

Details

This function will evaluate candidate mz pairs found within an `xcmsSet` object or any peak list by EvaluatePairsFromXCMSSet against the raw measurement data. This step is required to minimize redundancy and false positive results. It will allow to generate a number of informative quality control plots. As quite some input data is required for this function, please have a look in the vignette for an example. A special parameter in this function is `rolp` which can be set to 'non', 'pos', 'neg' or 'all'. It will influence the time performance of the function by determining how many peaks are effectively tested. If `rolp` is set to 'non', no overlapping peaks will be skipped, every individual mz-pair will be sequentially evaluated (slow but most informative). If it is set to 'pos' or 'neg', overlapping peaks (determined by experiment wide deconvolution) will not be tested additionally for positive or negative hits ('neg' is standard). If set to 'all' overlapping peaks will always be removed from the list of mz-pairs to be tested (fast).