Generate a table for the candidates obtained in EvaluateCandidateListAgainstRawData.

GenerateCandXLSX will produce a XLSX of a list containing test results objects.

Usage

GenerateCandXLSX(res_list = NULL, xlsx_file = NULL, rejected = FALSE)

Arguments

res_list

A list of result objects (each testing an individual mz pair).

xlsx_file

File name.

rejected

Logical. Prepare table of rejected candidates if TRUE.

Value

Candidate table as data.frame.

Details

Just a wrapper, to get the important information in a tabular layout.

Examples

# load evaluation result of example data and
# generate table within R (use parameter xlsx_file to write to file)
x <- GenerateCandXLSX(HiResTEC::res_list)
str(x)
#> 'data.frame':	1 obs. of  10 variables:
#>  $ ID             : num 1
#>  $ RT             : num 1027
#>  $ mz             : num 556
#>  $ dE             : num 41.1
#>  $ P              : num 2.86e-16
#>  $ Name           : chr ""
#>  $ row            : chr "23"
#>  $ BaseFormula    : chr "C5"
#>  $ Spectrum       : chr "540.2308:70229 541.2324:35446 542.2299:15238 543.2299:5300 554.2459:8944 555.2539:74536 556.2631:3312706 556.56"| __truncated__
#>  $ OverlappingLine: chr "10, 12, 9, 13, 14, 23, 25, 24, 3, 26, 4, 15, 33, 34, 5, 27, 35, 8"
x[,1:5]
#>   ID       RT      mz    dE            P
#> 1  1 1026.544 556.262 41.11 2.862525e-16